rename interfaces files to be snake case
