+++ src/main/python/interfaces/app/__init__.py
@@ -1,1 +1,1 @@
-from .appUI import AppUI
+from .app_ui import AppUI
