+++ src/main/python/interfaces/login/__init__.py
@@ -1,3 +1,3 @@
-from .loginUI import LoginUI
-from .loginPageUI import LoginPageUI
-from .loginCreateUI import CreatePageUI
+from .login_ui import LoginUI
+from .login_page_ui import LoginPageUI
+from .login_create_ui import CreatePageUI
