+++ src/main/python/interfaces/resources/resources_add_view_ui.py
